menu 'WARP Session Renew' instead of slide button for quick registration
This commit reverts back and replace these two commits:

* 179f42f - Update quick registration via button switch
* 1f92415 - Update quick registration via button switch

ABOUT QUICK REGISTRATION BY SLIDE BUTTON

The quick registration via button switch, it is a nice idea.
Unfortunately breaks the user experience (UX) in two fundamental ways.
First, when the users decide to unregister from the WARP network
service, they expect that the slide button will not work anymore
(consistency) unless they register into it again. Second, it quickly
and subtly reverses a previous critical decision that the users had
taken because for consistency with the previous "Registration Delete"
action, they will find them connected again into WARP network with a
new registration key. Which can be an unexpected result twice: not
being registered, not having a new session. Finally, it cuts just one
user operation: clicking on the WARP button which has the exact reason
to exist in order to revert the menu "Delete Registration" and create
a new session. Moreover, three seconds might not be enough or too much
depending on the bandwidth or the WARP quality of service and response
time.

A THIRD CASE CAN ARISE, A CORNER CASE BUT INTERESTING

Thinking about it, there is another case - mostly a corner case but
interesting - for which the unregistering action took place due a
crond script that suspended the WARP/Team session outside the local
office hours. For example, the registration is deleted at 6pm and
restored at 8am next day. In such a way the worker is out of the
organization VPN outside a specific window time and during bank
holidays. In this case, the quick registration with the slide button
creates an unexpected situation, again. Plus, it breaks their habit to
slide to connect only when the organisation connection is available.

Signed-off-by: Roberto A. Foglietta <roberto.foglietta@gmail.com>